feat: suppression bénévole, vérification rôle JWT, email confirmation mot de passe

+++ src/user/user.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, UseFilters, UseGuards, Req } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, UseFilters, UseGuards, Req, UnauthorizedException } from '@nestjs/common';
 import { UserService } from './user.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
@@ -17,7 +17,7 @@ export class UserController {
   @UseGuards(AuthGuard)
   @Get('me')
   async getMe(@Req() req) : Promise<User | null> {
-    return this.userService.findOne(req.user)
+    return this.userService.findOne(req.user.id)
   }
 
   @Get(':id')
@@ -28,22 +28,27 @@ export class UserController {
   @UseGuards(AuthGuard)
   @Patch()
   async update(@Req() req, @Body() updateUser: UpdateUserDto) : Promise<User> {
-    return this.userService.update(req.user, updateUser);
+    return this.userService.update(req.user.id, updateUser);
   }
 
   @UseGuards(AuthGuard)
   @Patch('onboarding')
   async updateSkillsFirstConnexion(@Req() req, @Body() updateUser: OnBoarding) : Promise<User> {
     
-    console.log(req.user);
+    console.log(req.user.id);
     
     
-    return this.userService.onboarding(req.user, updateUser);
+    return this.userService.onboarding(req.user.id, updateUser);
   }
 
 
+  @UseGuards(AuthGuard)
   @Delete(':id')
-  async remove(@Param('id') id: string) : Promise<void>{
+  async remove(@Req() req, @Param('id') id: string) : Promise<void>{
+
+    if(req.user.role !== 'admin' && req.user.id !== +id) {
+      throw new UnauthorizedException('Action non autorisée')
+    }
     return this.userService.remove(+id);
   }
 }
